fix: 포스트 api 통합 #134
